+++ auth/auth.module.go
@@ -1,11 +1,13 @@
 package auth
 
 import (
+	auth_lang "golang-api/auth/lang"
 	"golang-api/code"
 	"golang-api/database"
 	"golang-api/dotenv"
 	"golang-api/email"
 	"golang-api/gin"
+	"golang-api/lang"
 	"golang-api/log"
 	"golang-api/token"
 	"golang-api/user"
@@ -24,6 +26,7 @@ func NewAuthModule() *AuthModule {
 		Module: core.NewModule("AuthModule"),
 	}
 
+	module.AddModule(lang.Module())
 	module.AddModule(gin.Module())
 	module.AddModule(dotenv.Module())
 	module.AddModule(email.Module())
@@ -34,6 +37,10 @@ func NewAuthModule() *AuthModule {
 	module.AddModule(code.Module())
 	module.AddProvider(NewAuthService(module))
 	module.AddProvider(NewAuthController(module))
+
+	langService := module.Get("LangService").(lang.ILangService)
+	langService.AddStrings(auth_lang.EN_US, "en_US", "en_GB", "en_CA")
+	langService.AddStrings(auth_lang.FR_FR, "fr_FR", "fr_CA")
 
 	return module
 }