+++ auth/lang/en_US.go
@@ -1,0 +1,7 @@
+package auth_lang
+
+import "golang-api/lang"
+
+var EN_US = lang.LangString{
+	"invalid-credentials": "Invalid credentials",
+}