+++ auth/lang/fr_FR.go
@@ -1,0 +1,7 @@
+package auth_lang
+
+import "golang-api/lang"
+
+var FR_FR = lang.LangString{
+	"invalid-credentials": "Identifiants invalides",
+}