+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### Added
 
-- Add module, service, model, controller for `auth`, `user`, `code`, `database`, `dotenv`, `email`, `gin`, `log`, `query`, `token`, `websocket`
+- Add module, service, model, controller for `auth`, `user`, `code`, `database`, `dotenv`, `email`, `gin`, `log`, `query`, `token`, `websocket`, `media`, `lang`
 - Add unit test for `user`, `auth`, `database`, `dotenv`, `email`, `gin`, `log`, `media`
+- Add locale `en_US`, `fr_FR` for module `lang`
 - Add github action workflow for ci and deploy
 - Add docker and docker-compose setup